requirements.txt
# Core Kivy Framework
kivy>=2.3.0
kivymd>=1.2.0

# Cryptography and Security
pycryptodome>=3.20.0
cryptography>=41.0.0
ecdsa>=0.19.0
mnemonic>=0.21

# Blockchain and Consensus
hashlib-compat>=1.0.1
merklelib>=1.0.0

# Database and Storage
sqlite3-utils>=3.36.0
leveldb>=0.201

# HTTP Client and API Integration
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.27.0

# Async Support
asyncio-mqtt>=0.16.0
asyncio>=3.4.3

# Financial APIs
rapyd-python-sdk>=1.0.0
flutterwave-python>=3.0.0
nowpayments-api-python>=1.0.0

# Data Processing and Validation
pydantic>=2.5.0
marshmallow>=3.20.0
jsonschema>=4.20.0

# Threading and Concurrency
threading2>=0.1.2
concurrent-futures>=3.1.1

# Networking and P2P
xmlrpc-utils>=1.0.0
socket.io-client>=0.7.2

# GUI and Design
pillow>=10.1.0
matplotlib>=3.8.0
plotly>=5.17.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
colorama>=0.4.6
click>=8.1.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.11.0
flake8>=6.1.0

# Build and Deployment
buildozer>=1.5.0
python-for-android>=2024.1.21
cython>=3.0.0

# Mobile-specific
plyer>=2.1.0
android>=0.1.0

# Cross-platform compatibility
psutil>=5.9.0